test/install: add creation of missing rescue slot to fixture setup
Slot rescue-0 was described in the system.conf but not created in the
fixture setup for the install test. This is required for testing with
the upcoming mechanism that checks for missing slots and reports a
warning if a slot described in the system.conf is not found as a path in
the system.

Signed-off-by: Enrico Joerns <e.joerns@tu-bs.de>