manifest: allow specifying executed hooks per slot image
Adds a 'hooks' parameter to [images.<class>] manifest section. This
allows to specify the exact hooks executed for this image. The argument
ist a semicolon-separated list of per-slot-hooks.

Note that the name is secified without the implicit "slot-" prefix, i.e.
the hook executed when specifying "pre-install" will be
"slot-pre-install".

[image.rootfs]
...
filename = "my.img"
hooks = pre-install;post-install

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>