rauc-installer.xml: remove Completed signal
Using a completed signal to signal the result of a message call is an
anti-pattern. The method call and the signal emission do not have a
one-to-one relationship with each other. Thus, with one method call for
example, multiple 'Complete' signals could be emitted for which the
proper handling is not defined.

Returning from methods even after long-running operations is the right
design choice here.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>