service: add new D-Bus method 'mark'
Let the CLI call this method via D-Bus if RAUC has been configured to support it
and if not call it internally like before.

Adapt the tests to this new situation as well by starting the RAUC D-Bus service
on the system or the session bus according to whether they are running under
user mode linux or conventionally.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>