install/config_file: let load_slot_status() always return a valid status to simplify install logic
This allows to get rid of one layer of case differentiation in the install logic
while staying equivalent.

In case of a problem occuring in load_slot_status() instead of explicitely
modifying a GError** argument and returning directly do what the single caller
of the function did before: communicate the problem via g_message().

The impact on read_slot_status() is that it now acts on a pre-allocated data
structure and therefore has to free old memory before assigning newly allocated
components.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>