src/manifest: eliminate verify_manifest() function
Most things this function do are useless or buggy, thus it should be removed
and replaced by a simple call of verify_manifest_checksums().

* The function has a `signature` argument and handling for verifying
manifest signatures (only required for the obsolete network mode), but
it is never called with `signature` set to TRUE (thus the code sectio
can be removed, safeily)

* The function called check_compatible, which bypasses a possible custom
install-check handler and is redundant as the compatible will be checked
in launch_and_wait_{custom,default}_handler() again.

* The function also uses load_manifest() to load the manifest. This
requires to previously mount or unpack the bundle. Using the function
extract_manifest_from_bundle() is more generic for this.

Thus, what remains from this function is basically loading the bundle and
verifying checksums. This can be done separately with the advantage that
this also follows the convention for handling bundles.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>