src/context: split r_context_configure() into service and generic part
This is a preparation for dropping configuration parsing in cases where
this is not used which is essentially everything except the service daemon
and 'install' and 'status' in the non-D-Bus use case.

We need to split up as keyring_path and keyring-directory require extra
handling because they are used from bundle creating functions, too.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>