context: introduce loadconfig variable to load config only when required
This introduces 'loadconfig' variable to context.
This defaults to FALSE and must be set by command handlers that need to
have content of the RAUC config file available.

This is only the case for operations performed by the target and
normally only for the RAUC service daemon.

Only if RAUC is compiled without D-Bus support also 'status' and
'install' operations need to load the config file explicitly before.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>