src: store time as GDateTime in RaucSlotStatus to ease operations
Up to now the time was stored as a gchar string in RaucSlotStatus as it
was only recorded and never modified or used for any operation.

If we now aim to use this time to base further decisions on it, we
should be able to properly handle the time as time and not as a string.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>