src/main: fix newline handling to work for non-service case, too
In e463d776, basic progress bar support for RAUC was implemented.
For this, the prior line break in the g_print() in install_notify()
got replaced by adding distinct line breaks in the strings collected by
on_installer_changed().

However, for the non-service case, the messages do not go through D-Bus
but are processed from the message queue directly.
Because of that, there is no way to inject line breaks in the
non-service case and as a result of this the g_printed output does not
have any line breaks and is unreadable.

We fix this by reverting the original newline handling of
install_notify() and adapt the progress output string of
on_installer_changed() accordingly.

Site note: This will have the side-effect that, if progress bar output
is enabled, the lines above the progress line will now be printed with a
newline while the progress bar itself remains at the bottom of the
terminal output. This is basically the identical output one has without
progress bar, but with one extra line used for the bar.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>