src/main: no need to get primary slot for 'rauc status' mark methods
It is not required to call the primary boot slot determination method
(and thus communicate with the bootloader status) for other 'rauc
status' subcommands as 'mark-good' or 'mark-active'.

Thus we move this call to the conditional handling the bare 'rauc
status' call.

For non-service 'mark-xx' subcommand calls we only call
determine_slot_states(). This is what the corresponding D-Bus handling
method (r_on_handle_mark()) does too.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>