src/network.c: do not use g_clear_pointer() for fclose()
With some glib versions, using fclose as function argument for
g_clear_pointer() will cause a warning:

| /usr/include/glib-2.0/glib/gmem.h:132:31: warning: cast between incompatible function types from ‘int (*)(FILE *)’ to ‘void (*)(void *)’ [-Wcast-function-type]
|   132 |     GDestroyNotify _destroy = (GDestroyNotify) (destroy);                      \
|       |                               ^
| src/network.c:167:2: note: in expansion of macro ‘g_clear_pointer’
|   167 |  g_clear_pointer(&xfer.dl, fclose);
|       |  ^~~~~~~~~~~~~~~

This behavior was fixed, with a rework of g_clear_pointer type checking which
requires glib API version >= 2.58.

https://gitlab.gnome.org/GNOME/glib/-/merge_requests/177

Since we forced the API version to be 2.52, this will now also fail for builds
with more recent glib version.

This simply replaces the usage of g_clear_pointer by calling fclose()
explicitly.
This also allows to honor the return code and at least print a debug message if
closing failed.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>