src/bundle: do not use tmpdir for temporary encryption bundle
Using g_rename() to rename the temporary AES-encrypted bundle to the
final bundle requires them to be in the same file system.
This is normally not the case when using /tmp for storing the temporary
bundle.

Instead, create a random file name in the directory we also create our
final bundle in.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>