src: support verity bundles of 4096 bytes
Since bundle_verity_size of 0 is used in some places to signal an
uninitalized size, we use G_MAXUINT64 instead.

This is also safe as it is not 0 % 4096 and thus simply rejected by
current RAUC versions.

To not confuse when using 'rauc info', explicitly note that we do only
have a root hash.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>