src: add global state file
This introduces support for writing a global rauc state file, located
int the data-directory.

re-using global status is not optimal since:

a) information is quite slot-centric there, which is a different purpose
   than a global system view

b) using it for global state requires loading it early which would
   require major restructuring and then we still break per-slot status
   and/or force per-slot status to mount all partitions during startup.

Thus use a newly-introduced dedicated file in data-directory for storing
global status.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>