src/install: do not pass manifest custom handler args to all handlers
These args are meant to be passed to the full custom handler only, which
actually does not have much to do with the normal (system-defined)
handlers.

However, we have so far passed the handler args to all handlers.

Provide a better separation by obtaining the handler args in the custom
handler only and allow passing generic arguments to
launch_and_wait_handler().

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>