src/install: pass handler arguments as argv argument list
So far, the arguments to handlers were passed as a single string, which
may work or may not, depending on how the programm handles argument
parsing.
The more robust and correct way is to properly pass each individual
argument separately to g_subprocess.

For this purpose, we split up all parsed argument strings into (gchar**)
string arrays using g_shell_parse_argv() as we already do for other
arguments, too.

The launch_and_wait_handler() method is converted to take also take a
string array for arguments and all handing in between is adjusted
accordingly.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>