src/manifest: explicit error when using [handler] args without filename
Despite it was never documented or really intended, it was possible to use
the 'args' key in the manifest's '[handler]' section to pass arguments
to (system.conf-defined) handlers.
However, the actual intention of this key was to allow passing
additional arguments to the (manifest-defined) full custom handler.

Since we have just removed support for using the 'args' key for
system.conf-defined handlers, we now explicitly error-out in manifest
parsing in cases where 'args' was set but 'filename' (i.e. full custom
handler) was not.

This will prevent users from accidentally creating incompatible bundles.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>