src/global_state: create dummy slot for external boot
The external boot is still to be handled quite different than a normal
boot. The main reason is that one does not have any slot in state
ST_BOOTED.
To resolve this, we simply add a dummy readonly slot that gets the name
'external' and the external information as device.
This slot is set to ST_BOOTED.

In readable slot status output this will then also be visible like:

| === Slot States ===
|   [external] (_external_, virtual, booted)
|
| ...

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>