src/install: move compatible check handling out of launch_and_wait_*
The code is more or less the same for both cases and the small
differences in error handling are more 'grown' than by design.

Thus deduplicate the code here and make the flow a bit clearer.

This change has only little impact on the code flow: It moves the
compatible check before the respective context steps.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>