config_file: parse [artifacts.*] sections
These sections are used to configure artifact repositories. The repo
directories need to be created before starting RAUC. Creating them on
demand would make it harder to debug mount ordering problems during
startup.

Perhaps later, we could continue the start-up if a directory is missing
and mark the repo as missing and allow installation of bundles that
don't refer to the missing repos.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>