test/openssl-ca: add keyUsage to codeSigning configuration
Since v3.2.0, OpenSSL defines 'codesign' check purpose itself.

In the document, referenced there:

|  CA/Browser Forum,
|  Baseline Requirements for the Issuance and Management of
|  Publicly‐Trusted Code Signing Certificates, Version 3.0.0,
|  Section 7.1.2.3: Code signing and Timestamp Certificate

The 'keyUsage' paragraph in the corresponding PDF says:

| e. keyUsage
| This extension MUST be present and MUST be marked critical.
| The bit position for digitalSignature MUST be set. Bit positions for
| keyCertSign and cRLSign MUST NOT be set. All other bit positions SHOULD
| NOT be set

According to this, the behavior of OpenSSL to enforce EXFLAG_KUSAGE is
more correct than our implementation which did not enforce this.

The CI certificate created thus did not contain 'keyUsage' settings and
thus the CI aborts with OpenSSL 3.2.0 with:

| signature verification failed: Verify error: unsuitable certificate purpose

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>