src/signature: warn when using 'codesign' purpose without 'Key Usage'
According to the CAB Forum Requirements, noted in 'Baseline Requirements
for the Issuance and Management of Publicly‐Trusted Code Signing
Certificates, Version 3.0.0' [1] Section 7.1.2.3: 'Code signing and
Timestamp Certificate', the keyUsage extension 'MUST be present and MUST
be marked critical'.

The RAUC version of the check does not enforce this, but to be
compatible with the actual requirements and with the 'codesign' check as
implemented since OpenSSL 3.2.0, users should however define the 'Key
Usage' field for new certificates.

For this purpose, emit a warning when facing a certificate that does not
have this set.
Put the warning after having all other checks succeed to not irritate
with false-positive warnings.

[1] https://cabforum.org/wp-content/uploads/Baseline-Requirements-for-the-Issuance-and-Management-of-Code-Signing.v3.2.pdf

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>