src/install: map errors in check_bundle to R_INSTALL_ERROR_REJECTED
We use this code to differentiate a 'failed' and a 'rejected'
installation.

The expectation for a 'rejected' installation would be that some
pre-checks found the update to be inappropriate and refused to install
before actually modifying anything.

A 'failed' installation on the other hand implies that something went
wrong during actually performing the installation (i.e. writing data,
etc.).

Making this differentiation clearer would to provide a more consistent
high level error reporting and handling, for example in the event
logging code where this is already handled as different MESSAGE_IDs.