src/update_handler: add note on progress for adaptive fallback
With the previous addition of progress handling for adaptive mode hash
index generation, the progress may already be at a certain percentage
when an adaptive error occurs that triggers the manual copying fallback.
The manual copying will then reset the step progress to 0.

Since we allow the step progress to go backwards but ensure that high
level progress is only emitted when progress increases, this shouldn't
be an issue. The only observable behavior in this case is that the
overall progress may stall until the step percentage exceeds its
previous value again.
Since this case should happen rarely, it should be acceptable.

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>