src/update_handler: add handler for emmc-boot-linked
This adds an update handler for the 'emmc-boot-linked' slot type that
just copies the image to the respective emmc boot partition (without
touching the boot partition bit in extCSD).

The handler also handles migration from a different boot slot
configuration, like boot-emmc: If the target slot for the bootloader
update is the same as the currently used one, RAUC will backup the
current bootloader to the inactive boot partition first.

Signed-off-by: Lars Schmidt <l.schmidt@pengutronix.de>