test/emmc: separate emmc tests from write_slots
The emmc tests will get extended. As preparation the emmc tests
are moved to their own file. This separates the tests more clearly and
makes it easier to distinguish them from other tests. Especially the emmc
tests can not be executed in every environment as they need need
RAUC_TEST_EMMC in the os environment.

Signed-off-by: Lars Schmidt <l.schmidt@pengutronix.de>